face-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
deepface==0.0.79
python-dotenv==1.0.0
tf-keras
python-dotenv==1.0.0
